refactor: backend structure refactored
