+++ backend/pyproject.toml
@@ -20,6 +20,7 @@ passlib = "^1.7.4"
 bcrypt = "4.0.1"
 apscheduler = "^3.11.0"
 websockets = "^14.1"
+slowapi = "^0.1.9"
 
 
 [build-system]
