+++ backend/src/app/core/enums.py
