+++ backend/src/app/core/utils/security_utils.py
